GCS_Backend/bucket-emailer/requirements.txt
sendgrid
google-cloud-storage
google-cloud-error-reporting